requirements.txt
discord.py==2.3.2
asyncio==3.4.3
httpx==0.27.0
inflect==7.3.0